committed the release source

+++ packages/syncfusion_officechart/lib/src/chart/chart_text_area.dart
@@ -59,12 +59,12 @@ class ChartTextArea {
   }
 
   /// Gets font size.
-  int get size {
+  double get size {
     return _font.size;
   }
 
   /// Sets font size.
-  set size(int value) {
+  set size(double value) {
     _font.size = value;
   }
